Re-applied changes by clansley:
"The bignum code uses a "dirty character" of 254, which is ok as nobody
is going to be using 254 on the framebuffer; because it's the command
escape code for Matrix Orbitals.  The problem is that this "dirty
character" gets on to the wire and ends up sending random commands to
the LCD! There are also a few other small changes included.

Heres' a quick summary of the change:
  MtxOrb_draw_frame:
	- don't allow the dirty character on to the wire.
      - there's no need to keep sending custom chars (<8) they will
        automatically change when they are redefined.
	- share the "old" copy of the frame buffer with MtxOrb_flush_box
      (make it global)
	- remove commented out code - it's stored forever in CVS anyway.

  MtxOrb_flush_box:
	- <same as MtxOrb_draw_frame>
	- update the "old" copy of the frame buffer

  MtxOrb_init:
	- fill frame buffer with spaces, just to be safe.
	- malloc the "old" copy of the frame buffer and fill with dirty char

  MtxOrb_close:
	- free the "old" copy of the frame buffer

I've checked 0.5 and it doesn't seem to make sense to make this change
there (no dirty char and no MtxOrb_flush_box)."